fixed timer sometimes not starting
this is sort of a half fix but i cant find any better memory values. problem was that after launching the game, sometimes the chapter check would be at 0 until you entered a new chapter or rewound a chapter, which would mean the timer wouldn't start and the first split wouldn't work. the new address works the same as the old one, and again will sometimes be stuck at 0, however since wake-up is normally 0 with this one, it isn't as much of an issue.